Huge clean of unused code
