+++ gscore/cli/build.py
@@ -95,4 +95,4 @@ class Build:
 
     def __repr__(self):
 
-        return f"<Train> {self.name}"
+        return f"<Build> {self.name}"
