--- gscore/cli/common.py
@@ -1,8 +1,0 @@
-
-class PassArgs:
-
-    def __init__(self, args=dict()):
-        self._args = args
-
-    def __getattr__(self, key):
-        return self._args[key]
