+++ gscore/cli/gscore
@@ -50,5 +50,3 @@ if __name__ == '__main__':
         if len(sys.argv) == 1:
 
             parser.print_help()
-
-
