+++ gscore/combiner.py
@@ -162,65 +162,3 @@ class PrecursorExport(MutableMapping):
                 else:
 
                     writer.writerow(record)
-
-
-if __name__ == '__main__':
-    import glob
-
-    from gscore.parsers import osw
-    from gscore.parsers import queries
-
-    from gscore.fdr import GlobalDistribution
-
-    osw_files = glob.glob("/home/aaron/projects/aki/data/sample_specific/osw/*.osw")
-
-    global_protein_model = GlobalDistribution.load(
-        "/home/aaron/projects/aki/data/sample_specific/models/protein.model"
-    )
-
-    global_peptide_model = GlobalDistribution.load(
-        "/home/aaron/projects/aki/data/sample_specific/models/peptide.model"
-    )
-
-    from pathlib import Path
-
-    export = PrecursorExport(max_q_value=0.05)
-
-    for osw_file in osw_files[:10]:
-
-        print(f"{osw_file}")
-
-        sample_name = Path(osw_file).stem
-
-        export.add_sample(sample_name)
-
-        with osw.OSWFile(osw_file) as osw_conn:
-
-            precursors = osw_conn.parse_to_precursors(
-                query=queries.SelectPeakGroups.FETCH_ALL_SCORED_DATA
-            )
-
-            for precursor in precursors:
-
-                precursor_id = f"{precursor.modified_sequence}_{precursor.charge}"
-
-                if precursor_id not in export:
-
-                    export[precursor_id] = PrecursorExportRecord(
-                        modified_sequence=precursor.modified_sequence,
-                        charge=precursor.charge,
-                        decoy=precursor.decoy,
-                        protein_accession=precursor.protein_accession,
-                        protein_q_value=global_protein_model.get_q_value(precursor.protein_accession),
-                        peptide_q_value=global_peptide_model.get_q_value(precursor.modified_sequence)
-                    )
-
-                peakgroup = precursor.get_peakgroup(rank=1, key="q_value")
-
-                export[precursor_id].add_sample(
-                    sample_key=sample_name,
-                    peakgroup=peakgroup
-                )
-
-    export.write("/home/aaron/projects/aki/data/sample_specific/gscore/211207_sample_specific.tsv")
-
